# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: City Tour Italy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: experiences
    description: Tour experiences
  - name: auth
    description: Authentication
  - name: bookings
    description: Bookings
  - name: favorites
    description: Favorites
  - name: users
    description: User profile
  - name: reviews
    description: Reviews
  - name: coupons
    description: Coupons
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /experiences:
    get:
      operationId: listExperiences
      tags: [experiences]
      summary: List experiences with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: duration
          in: query
          schema:
            type: string
            enum: [short, medium, long]
        - name: minRating
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"

  /experiences/{id}:
    get:
      operationId: getExperience
      tags: [experiences]
      summary: Get experience by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Experience detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "404":
          description: Not found

  /experiences/stats/summary:
    get:
      operationId: getExperienceStats
      tags: [experiences]
      summary: Get marketplace stats summary
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceStats"

  /experiences/{id}/reviews:
    get:
      operationId: listExperienceReviews
      tags: [reviews]
      summary: List approved reviews for an experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of approved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for an experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          description: Unauthorized

  /coupons/validate:
    get:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized

  /bookings/my:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: List current user's bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized

  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List current user's favorites
      responses:
        "200":
          description: List of favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Favorite"
        "401":
          description: Unauthorized
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add experience to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"
        "401":
          description: Unauthorized

  /favorites/{experienceId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove experience from favorites
      parameters:
        - name: experienceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized

  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user's name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /admin/reviews/pending:
    get:
      operationId: listPendingReviews
      tags: [admin]
      summary: List reviews pending moderation
      responses:
        "200":
          description: List of pending reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "403":
          description: Forbidden

  /admin/reviews/{id}/approve:
    patch:
      operationId: approveReview
      tags: [admin]
      summary: Approve a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "403":
          description: Forbidden

  /admin/reviews/{id}/reply:
    patch:
      operationId: replyToReview
      tags: [admin]
      summary: Add a public reply to a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "200":
          description: Reply saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "403":
          description: Forbidden

  /admin/reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [admin]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden

  /admin/coupons:
    get:
      operationId: listCoupons
      tags: [admin]
      summary: List all coupons
      responses:
        "200":
          description: List of coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
        "403":
          description: Forbidden
    post:
      operationId: createCoupon
      tags: [admin]
      summary: Create a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "403":
          description: Forbidden

  /admin/coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [admin]
      summary: Update a coupon (activate/deactivate)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "403":
          description: Forbidden
    delete:
      operationId: deleteCoupon
      tags: [admin]
      summary: Delete a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Experience:
      type: object
      required: [id, title, shortDescription, longDescription, price, durationMinutes, category, rating, reviewCount, imageUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        shortDescription:
          type: string
        longDescription:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        category:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        imageUrl:
          type: string
        createdAt:
          type: string

    MarketplaceStats:
      type: object
      required: [totalExperiences, avgPrice, totalCategories, avgRating]
      properties:
        totalExperiences:
          type: integer
        avgPrice:
          type: number
        totalCategories:
          type: integer
        avgRating:
          type: number

    User:
      type: object
      required: [id, name, email, isAdmin, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    Booking:
      type: object
      required: [id, userId, experienceId, bookedDate, participants, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        experienceId:
          type: integer
        bookedDate:
          type: string
        participants:
          type: integer
        couponId:
          type: integer
          nullable: true
        discountPercent:
          type: integer
          nullable: true
        totalPrice:
          type: number
          nullable: true
        createdAt:
          type: string
        experience:
          $ref: "#/components/schemas/Experience"

    Favorite:
      type: object
      required: [id, userId, experienceId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        experienceId:
          type: integer
        createdAt:
          type: string
        experience:
          $ref: "#/components/schemas/Experience"

    Review:
      type: object
      required: [id, userId, experienceId, rating, comment, isApproved, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        experienceId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        reply:
          type: string
          nullable: true
        isApproved:
          type: boolean
        userName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string

    Coupon:
      type: object
      required: [id, code, discountPercent, validUntil, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        discountPercent:
          type: integer
        validUntil:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string

    CouponValidation:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        discountPercent:
          type: integer
          nullable: true
        code:
          type: string
          nullable: true
        message:
          type: string
          nullable: true

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    BookingInput:
      type: object
      required: [experienceId, bookedDate, participants]
      properties:
        experienceId:
          type: integer
        bookedDate:
          type: string
        participants:
          type: integer
          minimum: 1
          maximum: 10
        couponCode:
          type: string
          nullable: true

    FavoriteInput:
      type: object
      required: [experienceId]
      properties:
        experienceId:
          type: integer

    ProfileUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2

    ReviewInput:
      type: object
      required: [rating, comment]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 10

    ReplyInput:
      type: object
      required: [reply]
      properties:
        reply:
          type: string

    CouponInput:
      type: object
      required: [code, discountPercent, validUntil]
      properties:
        code:
          type: string
        discountPercent:
          type: integer
          minimum: 1
          maximum: 100
        validUntil:
          type: string
        isActive:
          type: boolean

    CouponUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        code:
          type: string
        discountPercent:
          type: integer
        validUntil:
          type: string