# lib/api-zod/src/generated/types/coupon.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * City Tour Italy API
 * OpenAPI spec version: 0.1.0
 */

export interface Coupon {
  id: number;
  code: string;
  discountPercent: number;
  validUntil: string;
  isActive: boolean;
  createdAt: string;
}